Implement authorization using Gates

+++ app/Http/Controllers/QuestionsController.php
@@ -64,7 +64,11 @@ class QuestionsController extends Controller
      */
     public function edit(Question $question)
     {
-        return view('questions.edit', compact('question'));
+        if (\Gate::allows('update-question',$question))
+        {
+            return view('questions.edit', compact('question'));
+        }
+        abort(403, "Access Denied") ;  
     }
 
     /**
@@ -76,6 +80,10 @@ class QuestionsController extends Controller
      */
     public function update(AskQuestionRequest $request, Question $question)
     {
+        if (\Gate::denies('update-question',$question))
+        {
+            abort(403, "Access Denied") ;
+        }
         $question->update($request->only('title', 'body'));
         return redirect('/questions')->with('success', "Your question has been updated.");
     }
@@ -88,6 +96,10 @@ class QuestionsController extends Controller
      */
     public function destroy(Question $question)
     {
+        if (\Gate::denies('delete-question',$question))
+        {
+            abort(403, "Access Denied") ;
+        }
         $question->delete();
         return redirect('/questions')->with('success', "Your question has been deleted.");
     }
